fix: delete downloaded chapter if only file exists.

+++ lib/features/downloads/services/delete_downloaded_chapter.dart
@@ -31,7 +31,10 @@ class DeleteDownloadedChapter {
         if (asset == null) return;
 
         if (asset.path != null) {
-          await File(asset.path!).delete();
+          final file = File(asset.path!);
+          if (await file.exists()) {
+            await File(asset.path!).delete();
+          }
         }
       },
     );
